refactor(cli/audio): split decoder submodule into file
